Add name attribute to metadata file.
In order to use Berkshelf with this cookbook, the metadata file must
correctly declare its name.  For this, we have two options:

1. Move to yum >= 2.0.0
2. Add the name attribute

The latter is much safer.  A more up-to-date copy of this cookbook lives in
the Altiscale repo and things that use Chef Server are already using it.
This is purely to add the ability to use test-kitchen to existing code.